temp-sensor/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-ahtx0==1.0.23
python-dotenv==1.0.1
requests==2.32.3
